mmretriever-api/requirements.txt
dataclasses-json
dataclasses
dashscope
elasticsearch>=8.0.0,<9.0.0
requests
pyyaml
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0